ai-log-analyzer-main/backend/requirements.txt
# --- Core backend ---
fastapi
uvicorn

# --- HTTP client (for remote LLM + webhooks) ---
requests

# --- Storage ---
minio

# --- Vector search ---
qdrant-client

# --- Embeddings ---
sentence-transformers
torch

# --- Utilities ---
python-dotenv
